pyproject.toml
[project]
name = "celine-flexibility-api"
version = "0.1.0"
description = "CELINE Flexibility API — commitment store for voluntary and automated load-shifting"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.12.0",
  "sqlalchemy>=2.0.0",
  "asyncpg>=0.30.0",
  "alembic>=1.13.0",
  "python-dotenv>=1.0",
  "PyJWT[cryptography]>=2.8.0",
  "httpx>=0.28.0",
  "celine-sdk>=1.5.0",
]

[tool.uv]

[dependency-groups]
dev = [
  "python-semantic-release>=10.5.3",
  "debugpy>=1.8.20",
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src/celine/flexibility"]
sources = {"src" = ""}

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]

build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""

branch = "main"
changelog_file = "CHANGELOG.md"
tag_format = "v{version}"
commit_parser = "conventional"

upload_to_release = false
upload_to_pypi = false
commit_version_number = false

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "up"]
parse_squash_commits = true
ignore_merge_commits = true
